extracted object spawn point logic
