+++ Assets/Scripts/Layers/Adapters/ObjImportAdapter.cs
@@ -209,7 +209,7 @@ namespace Netherlands3D.Twin
         
         public void AddLayerScriptToObj(GameObject parsedObj)
         {
-            var spawnPoint = GetSpawnPoint();
+            var spawnPoint = ObjectPlacementUtility.GetSpawnPoint();
 
             parsedObj.transform.position = spawnPoint;
             
@@ -217,18 +217,6 @@ namespace Netherlands3D.Twin
             parsedObj.AddComponent<MeshCollider>();
             objLayer.UI.Select();
             CreatedMoveableGameObject.Invoke(parsedObj);
-        }
-
-        private static Vector3 GetSpawnPoint()
-        {
-            var ray = Camera.main.ScreenPointToRay(new Vector2(Screen.width / 2, Screen.height / 2));
-            var plane = new Plane(Vector3.up, 0);
-            var intersect = plane.Raycast(ray, out float distance);
-            if (!intersect)
-                distance = 10f;
-
-            var spawnPoint = Camera.main.transform.position + Camera.main.transform.forward * distance;
-            return spawnPoint;
         }
     }
 }
