+++ Assets/Scripts/Layers/ObjectLibrary/ObjectLibraryButton.cs
@@ -33,7 +33,7 @@ namespace Netherlands3D.Twin
 
         private void CreateObject()
         {
-            var spawnPoint = GetSpawnPoint();
+            var spawnPoint = ObjectPlacementUtility.GetSpawnPoint();
             
             var newObject = Instantiate(prefab, spawnPoint, Quaternion.identity);
             newObject.name = prefab.name;
@@ -44,16 +44,5 @@ namespace Netherlands3D.Twin
             layerComponent.UI.Select();
         }
         
-        private static Vector3 GetSpawnPoint()
-        {
-            var ray = Camera.main.ScreenPointToRay(new Vector2(Screen.width / 2, Screen.height / 2));
-            var plane = new Plane(Vector3.up, 0);
-            var intersect = plane.Raycast(ray, out float distance);
-            if (!intersect)
-                distance = 10f;
-
-            var spawnPoint = Camera.main.transform.position + Camera.main.transform.forward * distance;
-            return spawnPoint;
-        }
     }
 }
