+++ Assets/Scripts/ObjectPlacementUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7c3540d6f00a4d6aa387f3b2ca053cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
